Substantial rework of how logging is setup in the API. No longer uses a .properties file. Instead, a Factory now has a static at class load time that spawns a NotesThread, which creates a brand new Session, gets the data directory from the Lotus API, then uses that to set up the logging behavior immediately.
In order to do this in an XPages context, we have to preempt the XPages Security Manager context.
So the OpenNTF Domino XSP plugin is now also an Equinox servlet, which means that it will be loaded as soon as the FIRST XPage is accessed on a Domino server. It also means that opening http://hostname/openntf_domino will report the version of the API on the target server.

This process also loads automatically when an Agent extends org.openntf.domino.AgentBase.

We should also add a manual entry point for the logging setup for use when people are loading from an individual NSF. And we should modify Factory.getDataPath() and similar functions to pre-cache these values when the logging is loaded so that some contextual information is available without constructing a Session, since that is so inherently delicate a process.