In theory, you can merge now, BUT...
You need to have m2e installed and configured
You need to have a Java JDK installed, not just Domino. But not Java 8 JDK only, because that breaks Guava. *sigh*
You'll have to manually configure a bunch of Maven repository artifacts. About 20 of them. I want to package these, but that's a task for tomorrow or the next day.
You might try one of the various p2 repository stories you can find documented on the web. I couldn't get them to work. In fact, the documentation I could find was simply horrible on the matter. It took 1/10th the time to just suck it up and wrap all the .jars in poms by hand that it did to even try one of the published repository strategies.
Maven documentation is so bad. It's clearly a very powerful build management system, but applying it to existing projects is like asking to get kicked in the crotch 100 times in a row.

We'll see if the strategy pays off going forward. I suspect it will since this will ultimately enable us to define profiles for builds that will make various versions and dependency trees exportable by a single command. The destination is great, but the journey is a rat-infested bilge bunk on a clipper with a cross-eyed navigator.